% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcRdiff_and_makeVolcanoPlot.R
\name{tcRDiff}
\alias{tcRDiff}
\title{tcrDiff: Performs differential motif analysis}
\usage{
tcRDiff(
  data,
  ident.1 = NULL,
  ident.2 = NULL,
  save = FALSE,
  expansion.score.threshold = Inf,
  fisher.score.threshold = Inf,
  ...
)
}
\arguments{
\item{data}{data.frame. GLIPH output .csv file}

\item{ident.1}{string. Identity class with a sample name of interest.}

\item{ident.2}{string. A second identity class for comparison.}

\item{save}{set save=TRUE to save results to a .csv file}

\item{expansion.score.threshold}{numeric. To threshold enrichment of clonal expansion within the cluster.}

\item{fisher.score.threshold}{numeric. A threshold for the p-value obtained by performing the Fisher's exact test with a contingency table containing unique_cdr3_sample, unique_cdr3_ref, the number of remaining sample sequences and the number of remaining reference sequences.The score reports the probability that a random sample of the same size as the sample set but into the reference set (i.e. naive repertoire) would generate an enrichment of the given motif at least as high as has been observed in the sample set.}
}
\description{
tcRdiff helps with an in-depth characterization of T-cell repertoires. It uses the output from GLIPH2 to distinguishing clonally expanded motifs. Clonally expanded and enriched motifs are identified by comparing the summed contribution scores of samples from each cluster (summed template frequency) using exact poisson statistics. Enriched motifs' log2-fold change and p-values can be visualized as volcano plots to identify expanded TCRs. The function outputs a data.table that can be visualized using the makeVolcanoPlot function.
}
\examples{

# Read GLIPH output with 'sample_name' column (with only sample information)
data <- read.csv("36604540_GLIPH2_Output.csv")

# Use function
df <- tcRDiff(data, ident.1 = "MtbLys",ident.2 = "D360", save = F, expansion.score.threshold = 0.05, fisher.score.threshold = 0.05)

}
